requirements.txt
# =============================================================================
# Adizon Trooper Worker - Python Dependencies
# Compute-intensive microservice for document processing & graph extraction
# =============================================================================

# -----------------------------------------------------------------------------
# Web Framework
# -----------------------------------------------------------------------------
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18

# -----------------------------------------------------------------------------
# Database & ORM
# -----------------------------------------------------------------------------
sqlalchemy[asyncio]>=2.0.36
psycopg[binary,pool]>=3.2.3
pgvector>=0.3.6

# -----------------------------------------------------------------------------
# Neo4j Graph Database
# -----------------------------------------------------------------------------
neo4j>=5.26.0

# -----------------------------------------------------------------------------
# LangChain & LangGraph
# -----------------------------------------------------------------------------
langchain>=0.3.13
langchain-openai>=0.2.14
langchain-community>=0.3.13
langchain-text-splitters>=0.3.4
langgraph>=0.2.60
langchain-postgres>=0.0.12

# -----------------------------------------------------------------------------
# Document Processing (PDF, DOCX, OCR)
# -----------------------------------------------------------------------------
pypdf>=5.1.0
docx2txt>=0.8
tiktoken>=0.8.0
unstructured>=0.16.0
unstructured[pdf]>=0.16.0
pdf2image>=1.17.0
pytesseract>=0.3.10
pillow>=10.0.0

# -----------------------------------------------------------------------------
# Object Storage (MinIO/S3)
# -----------------------------------------------------------------------------
boto3>=1.35.0

# -----------------------------------------------------------------------------
# Configuration & Utilities
# -----------------------------------------------------------------------------
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
aiofiles>=24.1.0
pyyaml>=6.0.0
httpx>=0.28.0
